Work on SideBar for Documentation
Animations work quite well, some animations are a bit funny when
transitioning screen sizes, but that rarely happens so it should be ok.
Built-In Functions page shows extra lines in the code, which is weird
and will have to be looked at.
Need a hit menu button inside the menu for smaller screens.